split memory management into separate files - no functional changes
