+++ CMakeLists.txt
@@ -980,6 +980,10 @@ set(LIBNETDATA_FILES
         src/libnetdata/locks/benchmark-rw.c
         src/libnetdata/locks/benchmark-rw.h
         src/libnetdata/log/nd_log-libunwind.c
+        src/libnetdata/memory/nd-mallocz.c
+        src/libnetdata/memory/nd-mallocz.h
+        src/libnetdata/memory/nd-mmap.c
+        src/libnetdata/memory/nd-mmap.h
 )
 
 set(LIBH2O_FILES
