+++ src/libnetdata/libnetdata.c
@@ -6,472 +6,9 @@
 #define size_t_atomic_count(op, var, size) __atomic_## op ##_fetch(&(var), size, __ATOMIC_RELAXED)
 #define size_t_atomic_bytes(op, var, size) __atomic_## op ##_fetch(&(var), ((size) % MALLOC_ALIGNMENT)?((size) + MALLOC_ALIGNMENT - ((size) % MALLOC_ALIGNMENT)):(size), __ATOMIC_RELAXED)
 
-#if !defined(MADV_DONTFORK)
-#define MADV_DONTFORK 0
-#endif
-
-#if !defined(O_NOATIME)
-#define O_NOATIME 0
-#endif
-
 struct rlimit rlimit_nofile = { .rlim_cur = 1024, .rlim_max = 1024 };
 
-#if defined(MADV_MERGEABLE)
-int enable_ksm = CONFIG_BOOLEAN_AUTO;
-#else
-int enable_ksm = 0;
-#endif
-
 volatile sig_atomic_t netdata_exit = 0;
-
-// ----------------------------------------------------------------------------
-// memory allocation functions that handle failures
-
-// although netdata does not use memory allocations too often (netdata tries to
-// maintain its memory footprint stable during runtime, i.e. all buffers are
-// allocated during initialization and are adapted to current use throughout
-// its lifetime), these can be used to override the default system allocation
-// routines.
-
-#ifdef NETDATA_TRACE_ALLOCATIONS
-#warning NETDATA_TRACE_ALLOCATIONS ENABLED
-#include "Judy.h"
-
-#if defined(HAVE_DLSYM) && defined(ENABLE_DLSYM)
-#include <dlfcn.h>
-
-typedef void (*libc_function_t)(void);
-
-static void *malloc_first_run(size_t size);
-static void *(*libc_malloc)(size_t) = malloc_first_run;
-
-static void *calloc_first_run(size_t n, size_t size);
-static void *(*libc_calloc)(size_t, size_t) = calloc_first_run;
-
-static void *realloc_first_run(void *ptr, size_t size);
-static void *(*libc_realloc)(void *, size_t) = realloc_first_run;
-
-static void free_first_run(void *ptr);
-static void (*libc_free)(void *) = free_first_run;
-
-static char *strdup_first_run(const char *s);
-static char *(*libc_strdup)(const char *) = strdup_first_run;
-
-static char *strndup_first_run(const char *s, size_t len);
-static char *(*libc_strndup)(const char *, size_t) = strndup_first_run;
-
-static size_t malloc_usable_size_first_run(void *ptr);
-#ifdef HAVE_MALLOC_USABLE_SIZE
-static size_t (*libc_malloc_usable_size)(void *) = malloc_usable_size_first_run;
-#else
-static size_t (*libc_malloc_usable_size)(void *) = NULL;
-#endif
-
-static void link_system_library_function(libc_function_t *func_pptr, const char *name, bool required) {
-    *func_pptr = dlsym(RTLD_NEXT, name);
-    if(!*func_pptr && required) {
-        fprintf(stderr, "FATAL: Cannot find system's %s() function.\n", name);
-        abort();
-    }
-}
-
-static void *malloc_first_run(size_t size) {
-    link_system_library_function((libc_function_t *) &libc_malloc, "malloc", true);
-    return libc_malloc(size);
-}
-
-static void *calloc_first_run(size_t n, size_t size) {
-    link_system_library_function((libc_function_t *) &libc_calloc, "calloc", true);
-    return libc_calloc(n, size);
-}
-
-static void *realloc_first_run(void *ptr, size_t size) {
-    link_system_library_function((libc_function_t *) &libc_realloc, "realloc", true);
-    return libc_realloc(ptr, size);
-}
-
-static void free_first_run(void *ptr) {
-    link_system_library_function((libc_function_t *) &libc_free, "free", true);
-    libc_free(ptr);
-}
-
-static char *strdup_first_run(const char *s) {
-    link_system_library_function((libc_function_t *) &libc_strdup, "strdup", true);
-    return libc_strdup(s);
-}
-
-static char *strndup_first_run(const char *s, size_t len) {
-    link_system_library_function((libc_function_t *) &libc_strndup, "strndup", true);
-    return libc_strndup(s, len);
-}
-
-static size_t malloc_usable_size_first_run(void *ptr) {
-    link_system_library_function((libc_function_t *) &libc_malloc_usable_size, "malloc_usable_size", false);
-
-    if(libc_malloc_usable_size)
-        return libc_malloc_usable_size(ptr);
-    else
-        return 0;
-}
-
-void *malloc(size_t size) {
-    return mallocz(size);
-}
-
-void *calloc(size_t n, size_t size) {
-    return callocz(n, size);
-}
-
-void *realloc(void *ptr, size_t size) {
-    return reallocz(ptr, size);
-}
-
-void *reallocarray(void *ptr, size_t n, size_t size) {
-    return reallocz(ptr, n * size);
-}
-
-void free(void *ptr) {
-    freez(ptr);
-}
-
-char *strdup(const char *s) {
-    return strdupz(s);
-}
-
-char *strndup(const char *s, size_t len) {
-    return strndupz(s, len);
-}
-
-size_t malloc_usable_size(void *ptr) {
-    return mallocz_usable_size(ptr);
-}
-#else // !HAVE_DLSYM
-
-static void *(*libc_malloc)(size_t) = malloc;
-static void *(*libc_calloc)(size_t, size_t) = calloc;
-static void *(*libc_realloc)(void *, size_t) = realloc;
-static void (*libc_free)(void *) = free;
-
-#ifdef HAVE_MALLOC_USABLE_SIZE
-static size_t (*libc_malloc_usable_size)(void *) = malloc_usable_size;
-#else
-static size_t (*libc_malloc_usable_size)(void *) = NULL;
-#endif
-
-#endif // HAVE_DLSYM
-
-
-void posix_memfree(void *ptr) {
-    libc_free(ptr);
-}
-
-struct malloc_header_signature {
-    uint32_t magic;
-    uint32_t size;
-    struct malloc_trace *trace;
-};
-
-struct malloc_header {
-    struct malloc_header_signature signature;
-    uint8_t padding[(sizeof(struct malloc_header_signature) % MALLOC_ALIGNMENT) ? MALLOC_ALIGNMENT - (sizeof(struct malloc_header_signature) % MALLOC_ALIGNMENT) : 0];
-    uint8_t data[];
-};
-
-static size_t malloc_header_size = sizeof(struct malloc_header);
-
-int malloc_trace_compare(void *A, void *B) {
-    struct malloc_trace *a = A;
-    struct malloc_trace *b = B;
-    return strcmp(a->function, b->function);
-}
-
-static avl_tree_lock malloc_trace_index = {
-    .avl_tree = {
-        .root = NULL,
-        .compar = malloc_trace_compare},
-    .rwlock = AVL_LOCK_INITIALIZER
-};
-
-int malloc_trace_walkthrough(int (*callback)(void *item, void *data), void *data) {
-    return avl_traverse_lock(&malloc_trace_index, callback, data);
-}
-
-NEVERNULL WARNUNUSED
-static struct malloc_trace *malloc_trace_find_or_create(const char *file, const char *function, size_t line) {
-    struct malloc_trace tmp = {
-        .line = line,
-        .function = function,
-        .file = file,
-    };
-
-    struct malloc_trace *t = (struct malloc_trace *)avl_search_lock(&malloc_trace_index, (avl_t *)&tmp);
-    if(!t) {
-        t = libc_calloc(1, sizeof(struct malloc_trace));
-        if(!t) fatal("No memory");
-        t->line = line;
-        t->function = function;
-        t->file = file;
-
-        struct malloc_trace *t2 = (struct malloc_trace *)avl_insert_lock(&malloc_trace_index, (avl_t *)t);
-        if(t2 != t)
-            free(t);
-
-        t = t2;
-    }
-
-    if(!t)
-        fatal("Cannot insert to AVL");
-
-    return t;
-}
-
-void malloc_trace_mmap(size_t size) {
-    struct malloc_trace *p = malloc_trace_find_or_create("unknown", "netdata_mmap", 1);
-    size_t_atomic_count(add, p->mmap_calls, 1);
-    size_t_atomic_count(add, p->allocations, 1);
-    size_t_atomic_bytes(add, p->bytes, size);
-}
-
-void malloc_trace_munmap(size_t size) {
-    struct malloc_trace *p = malloc_trace_find_or_create("unknown", "netdata_mmap", 1);
-    size_t_atomic_count(add, p->munmap_calls, 1);
-    size_t_atomic_count(sub, p->allocations, 1);
-    size_t_atomic_bytes(sub, p->bytes, size);
-}
-
-void *mallocz_int(size_t size, const char *file, const char *function, size_t line) {
-    struct malloc_trace *p = malloc_trace_find_or_create(file, function, line);
-
-    size_t_atomic_count(add, p->malloc_calls, 1);
-    size_t_atomic_count(add, p->allocations, 1);
-    size_t_atomic_bytes(add, p->bytes, size);
-
-    struct malloc_header *t = (struct malloc_header *)libc_malloc(malloc_header_size + size);
-    if (unlikely(!t)) fatal("mallocz() cannot allocate %zu bytes of memory (%zu with header).", size, malloc_header_size + size);
-    t->signature.magic = 0x0BADCAFE;
-    t->signature.trace = p;
-    t->signature.size = size;
-
-#ifdef NETDATA_INTERNAL_CHECKS
-    for(ssize_t i = 0; i < (ssize_t)sizeof(t->padding) ;i++) // signed to avoid compiler warning when zero-padded
-        t->padding[i] = 0xFF;
-#endif
-
-    return (void *)&t->data;
-}
-
-void *callocz_int(size_t nmemb, size_t size, const char *file, const char *function, size_t line) {
-    struct malloc_trace *p = malloc_trace_find_or_create(file, function, line);
-    size = nmemb * size;
-
-    size_t_atomic_count(add, p->calloc_calls, 1);
-    size_t_atomic_count(add, p->allocations, 1);
-    size_t_atomic_bytes(add, p->bytes, size);
-
-    struct malloc_header *t = (struct malloc_header *)libc_calloc(1, malloc_header_size + size);
-    if (unlikely(!t)) fatal("mallocz() cannot allocate %zu bytes of memory (%zu with header).", size, malloc_header_size + size);
-    t->signature.magic = 0x0BADCAFE;
-    t->signature.trace = p;
-    t->signature.size = size;
-
-#ifdef NETDATA_INTERNAL_CHECKS
-    for(ssize_t i = 0; i < (ssize_t)sizeof(t->padding) ;i++) // signed to avoid compiler warning when zero-padded
-        t->padding[i] = 0xFF;
-#endif
-
-    return &t->data;
-}
-
-char *strdupz_int(const char *s, const char *file, const char *function, size_t line) {
-    struct malloc_trace *p = malloc_trace_find_or_create(file, function, line);
-    size_t size = strlen(s) + 1;
-
-    size_t_atomic_count(add, p->strdup_calls, 1);
-    size_t_atomic_count(add, p->allocations, 1);
-    size_t_atomic_bytes(add, p->bytes, size);
-
-    struct malloc_header *t = (struct malloc_header *)libc_malloc(malloc_header_size + size);
-    if (unlikely(!t)) fatal("strdupz() cannot allocate %zu bytes of memory (%zu with header).", size, malloc_header_size + size);
-    t->signature.magic = 0x0BADCAFE;
-    t->signature.trace = p;
-    t->signature.size = size;
-
-#ifdef NETDATA_INTERNAL_CHECKS
-    for(ssize_t i = 0; i < (ssize_t)sizeof(t->padding) ;i++) // signed to avoid compiler warning when zero-padded
-        t->padding[i] = 0xFF;
-#endif
-
-    memcpy(&t->data, s, size);
-    return (char *)&t->data;
-}
-
-char *strndupz_int(const char *s, size_t len, const char *file, const char *function, size_t line) {
-    struct malloc_trace *p = malloc_trace_find_or_create(file, function, line);
-    size_t size = len + 1;
-
-    size_t_atomic_count(add, p->strdup_calls, 1);
-    size_t_atomic_count(add, p->allocations, 1);
-    size_t_atomic_bytes(add, p->bytes, size);
-
-    struct malloc_header *t = (struct malloc_header *)libc_malloc(malloc_header_size + size);
-    if (unlikely(!t)) fatal("strndupz() cannot allocate %zu bytes of memory (%zu with header).", size, malloc_header_size + size);
-    t->signature.magic = 0x0BADCAFE;
-    t->signature.trace = p;
-    t->signature.size = size;
-
-#ifdef NETDATA_INTERNAL_CHECKS
-    for(ssize_t i = 0; i < (ssize_t)sizeof(t->padding) ;i++) // signed to avoid compiler warning when zero-padded
-        t->padding[i] = 0xFF;
-#endif
-
-    memcpy(&t->data, s, size);
-    t->data[len] = '\0';
-    return (char *)&t->data;
-}
-
-static struct malloc_header *malloc_get_header(void *ptr, const char *caller, const char *file, const char *function, size_t line) {
-    uint8_t *ret = (uint8_t *)ptr - malloc_header_size;
-    struct malloc_header *t = (struct malloc_header *)ret;
-
-    if(t->signature.magic != 0x0BADCAFE) {
-        netdata_log_error("pointer %p is not our pointer (called %s() from %zu@%s, %s()).", ptr, caller, line, file, function);
-        return NULL;
-    }
-
-    return t;
-}
-
-void *reallocz_int(void *ptr, size_t size, const char *file, const char *function, size_t line) {
-    if(!ptr) return mallocz_int(size, file, function, line);
-
-    struct malloc_header *t = malloc_get_header(ptr, __FUNCTION__, file, function, line);
-    if(!t)
-        return libc_realloc(ptr, size);
-
-    if(t->signature.size == size) return ptr;
-    size_t_atomic_count(add, t->signature.trace->free_calls, 1);
-    size_t_atomic_count(sub, t->signature.trace->allocations, 1);
-    size_t_atomic_bytes(sub, t->signature.trace->bytes, t->signature.size);
-
-    struct malloc_trace *p = malloc_trace_find_or_create(file, function, line);
-    size_t_atomic_count(add, p->realloc_calls, 1);
-    size_t_atomic_count(add, p->allocations, 1);
-    size_t_atomic_bytes(add, p->bytes, size);
-
-    t = (struct malloc_header *)libc_realloc(t, malloc_header_size + size);
-    if (unlikely(!t)) fatal("reallocz() cannot allocate %zu bytes of memory (%zu with header).", size, malloc_header_size + size);
-    t->signature.magic = 0x0BADCAFE;
-    t->signature.trace = p;
-    t->signature.size = size;
-
-#ifdef NETDATA_INTERNAL_CHECKS
-    for(ssize_t i = 0; i < (ssize_t)sizeof(t->padding) ;i++) // signed to avoid compiler warning when zero-padded
-        t->padding[i] = 0xFF;
-#endif
-
-    return (void *)&t->data;
-}
-
-size_t mallocz_usable_size_int(void *ptr, const char *file, const char *function, size_t line) {
-    if(unlikely(!ptr)) return 0;
-
-    struct malloc_header *t = malloc_get_header(ptr, __FUNCTION__, file, function, line);
-    if(!t) {
-        if(libc_malloc_usable_size)
-            return libc_malloc_usable_size(ptr);
-        else
-            return 0;
-    }
-
-    return t->signature.size;
-}
-
-void freez_int(void *ptr, const char *file, const char *function, size_t line) {
-    if(unlikely(!ptr)) return;
-
-    struct malloc_header *t = malloc_get_header(ptr, __FUNCTION__, file, function, line);
-    if(!t) {
-        libc_free(ptr);
-        return;
-    }
-
-    size_t_atomic_count(add, t->signature.trace->free_calls, 1);
-    size_t_atomic_count(sub, t->signature.trace->allocations, 1);
-    size_t_atomic_bytes(sub, t->signature.trace->bytes, t->signature.size);
-
-#ifdef NETDATA_INTERNAL_CHECKS
-    // it should crash if it is used after freeing it
-    memset(t, 0, malloc_header_size + t->signature.size);
-#endif
-
-    libc_free(t);
-}
-#else
-
-char *strdupz(const char *s) {
-    char *t = strdup(s);
-    if (unlikely(!t)) {
-        OS_SYSTEM_MEMORY sm = os_last_reported_system_memory();
-        fatal("Cannot strdup() string '%s' (system memory available bytes: %lu)", s, sm.ram_available_bytes);
-    }
-    return t;
-}
-
-char *strndupz(const char *s, size_t len) {
-    char *t = strndup(s, len);
-    if (unlikely(!t)) {
-        OS_SYSTEM_MEMORY sm = os_last_reported_system_memory();
-        fatal("Cannot strndup() string '%s' of len %zu (system memory available bytes: %lu)", s, len, sm.ram_available_bytes);
-    }
-    return t;
-}
-
-// If ptr is NULL, no operation is performed.
-void freez(void *ptr) {
-    if(likely(ptr)) free(ptr);
-}
-
-void *mallocz(size_t size) {
-    void *p = malloc(size);
-    if (unlikely(!p)) {
-        OS_SYSTEM_MEMORY sm = os_last_reported_system_memory();
-        fatal("Cannot allocate %zu bytes of memory (system memory available bytes: %lu)", size, sm.ram_available_bytes);
-    }
-    return p;
-}
-
-void *callocz(size_t nmemb, size_t size) {
-    void *p = calloc(nmemb, size);
-    if (unlikely(!p)) {
-        OS_SYSTEM_MEMORY sm = os_last_reported_system_memory();
-        fatal("Cannot allocate %zu bytes of memory (system memory available bytes: %lu)", nmemb * size, sm.ram_available_bytes);
-    }
-    return p;
-}
-
-void *reallocz(void *ptr, size_t size) {
-    void *p = realloc(ptr, size);
-    if (unlikely(!p)) {
-        OS_SYSTEM_MEMORY sm = os_last_reported_system_memory();
-        fatal("Cannot re-allocate memory to %zu bytes. (system memory available bytes: %lu)", size, sm.ram_available_bytes);
-    }
-    return p;
-}
-
-void posix_memfree(void *ptr) {
-    free(ptr);
-}
-#endif
-
-void mallocz_release_as_much_memory_to_the_system(void) {
-#if defined(HAVE_C_MALLOC_TRIM)
-    static SPINLOCK spinlock = SPINLOCK_INITIALIZER;
-    if(spinlock_trylock(&spinlock)) {
-        malloc_trim(0);
-        spinlock_unlock(&spinlock);
-    }
-#endif
-}
 
 // --------------------------------------------------------------------------------------------------------------------
 
@@ -504,188 +41,6 @@ void json_fix_string(char *s) {
         else
             s++;
     }
-}
-
-static int memory_file_open(const char *filename, size_t size) {
-    // netdata_log_info("memory_file_open('%s', %zu", filename, size);
-
-    int fd = open(filename, O_RDWR | O_CREAT | O_NOATIME | O_CLOEXEC, 0664);
-    if (fd != -1) {
-        if (lseek(fd, size, SEEK_SET) == (off_t) size) {
-            if (write(fd, "", 1) == 1) {
-                if (ftruncate(fd, size))
-                    netdata_log_error("Cannot truncate file '%s' to size %zu. Will use the larger file.", filename, size);
-            }
-            else
-                netdata_log_error("Cannot write to file '%s' at position %zu.", filename, size);
-        }
-        else
-            netdata_log_error("Cannot seek file '%s' to size %zu.", filename, size);
-    }
-    else
-        netdata_log_error("Cannot create/open file '%s'.", filename);
-
-    return fd;
-}
-
-inline int madvise_sequential(void *mem, size_t len) {
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_SEQUENTIAL);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_SEQUENTIAL) of size %zu, failed.", len);
-    return ret;
-}
-
-inline int madvise_random(void *mem, size_t len) {
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_RANDOM);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_RANDOM) of size %zu, failed.", len);
-    return ret;
-}
-
-inline int madvise_dontfork(void *mem, size_t len) {
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_DONTFORK);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_DONTFORK) of size %zu, failed.", len);
-    return ret;
-}
-
-inline int madvise_willneed(void *mem, size_t len) {
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_WILLNEED);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_WILLNEED) of size %zu, failed.", len);
-    return ret;
-}
-
-inline int madvise_dontneed(void *mem, size_t len) {
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_DONTNEED);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_DONTNEED) of size %zu, failed.", len);
-    return ret;
-}
-
-inline int madvise_dontdump(void *mem __maybe_unused, size_t len __maybe_unused) {
-#if __linux__
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_DONTDUMP);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_DONTDUMP) of size %zu, failed.", len);
-    return ret;
-#else
-    return 0;
-#endif
-}
-
-inline int madvise_mergeable(void *mem __maybe_unused, size_t len __maybe_unused) {
-#ifdef MADV_MERGEABLE
-    static int logger = 1;
-    int ret = madvise(mem, len, MADV_MERGEABLE);
-
-    if (ret != 0 && logger-- > 0)
-        netdata_log_error("madvise(MADV_MERGEABLE) of size %zu, failed.", len);
-    return ret;
-#else
-    return 0;
-#endif
-}
-
-size_t netdata_mmap_count = 0;
-
-void *netdata_mmap(const char *filename, size_t size, int flags, int ksm, bool read_only, bool dont_dump, int *open_fd)
-{
-    // netdata_log_info("netdata_mmap('%s', %zu", filename, size);
-
-    // MAP_SHARED is used in memory mode map
-    // MAP_PRIVATE is used in memory mode ram and save
-
-    if(unlikely(!(flags & MAP_SHARED) && !(flags & MAP_PRIVATE)))
-        fatal("Neither MAP_SHARED or MAP_PRIVATE were given to netdata_mmap()");
-
-    if(unlikely((flags & MAP_SHARED) && (flags & MAP_PRIVATE)))
-        fatal("Both MAP_SHARED and MAP_PRIVATE were given to netdata_mmap()");
-
-    if(unlikely((flags & MAP_SHARED) && (!filename || !*filename)))
-        fatal("MAP_SHARED requested, without a filename to netdata_mmap()");
-
-    // don't enable ksm is the global setting is disabled
-    if(unlikely(!enable_ksm)) ksm = 0;
-
-    // KSM only merges anonymous (private) pages, never pagecache (file) pages
-    // but MAP_PRIVATE without MAP_ANONYMOUS it fails too, so we need it always
-    if((flags & MAP_PRIVATE)) flags |= MAP_ANONYMOUS;
-
-    int fd = -1;
-    void *mem = MAP_FAILED;
-
-    if(filename && *filename) {
-        // open/create the file to be used
-        fd = memory_file_open(filename, size);
-        if(fd == -1) goto cleanup;
-    }
-
-    int fd_for_mmap = fd;
-    if(fd != -1 && (flags & MAP_PRIVATE)) {
-        // this is MAP_PRIVATE allocation
-        // no need for mmap() to use our fd
-        // we will copy the file into the memory allocated
-        fd_for_mmap = -1;
-    }
-
-    mem = mmap(NULL, size, read_only ? PROT_READ : PROT_READ | PROT_WRITE, flags, fd_for_mmap, 0);
-    if (mem != MAP_FAILED) {
-
-#ifdef NETDATA_TRACE_ALLOCATIONS
-        malloc_trace_mmap(size);
-#endif
-
-        // if we have a file open, but we didn't give it to mmap(),
-        // we have to read the file into the memory block we allocated
-        if(fd != -1 && fd_for_mmap == -1) {
-            if (lseek(fd, 0, SEEK_SET) == 0) {
-                if (read(fd, mem, size) != (ssize_t) size)
-                    netdata_log_info("Cannot read from file '%s'", filename);
-            }
-            else netdata_log_info("Cannot seek to beginning of file '%s'.", filename);
-        }
-
-        // madvise_sequential(mem, size);
-        // madvise_dontfork(mem, size); // aral is initialized before we daemonize
-        if(dont_dump) madvise_dontdump(mem, size);
-        // if(flags & MAP_SHARED) madvise_willneed(mem, size);
-        if(ksm) madvise_mergeable(mem, size);
-    }
-
-cleanup:
-    if(fd != -1) {
-        if (open_fd)
-            *open_fd = fd;
-        else
-            close(fd);
-    }
-    if(mem == MAP_FAILED) return NULL;
-    __atomic_add_fetch(&netdata_mmap_count, 1, __ATOMIC_RELAXED);
-    errno_clear();
-    return mem;
-}
-
-int netdata_munmap(void *ptr, size_t size) {
-#ifdef NETDATA_TRACE_ALLOCATIONS
-    malloc_trace_munmap(size);
-#endif
-    int rc = munmap(ptr, size);
-    if(rc == 0)
-        __atomic_sub_fetch(&netdata_mmap_count, 1, __ATOMIC_RELAXED);
-    return rc;
 }
 
 char *fgets_trim_len(char *buf, size_t buf_size, FILE *fp, size_t *len) {
