+++ src/libnetdata/libnetdata.h
@@ -8,6 +8,8 @@ extern "C" {
 # endif
 
 #include "common.h"
+#include "memory/nd-mallocz.h"
+#include "memory/nd-mmap.h"
 #include "log/nd_log-fatal.h"
 #include "atomics/atomics.h"
 
@@ -23,60 +25,14 @@ extern "C" {
 #include "storage-point.h"
 #include "paths/paths.h"
 
-int madvise_sequential(void *mem, size_t len);
-int madvise_random(void *mem, size_t len);
-int madvise_dontfork(void *mem, size_t len);
-int madvise_willneed(void *mem, size_t len);
-int madvise_dontneed(void *mem, size_t len);
-int madvise_dontdump(void *mem, size_t len);
-int madvise_mergeable(void *mem, size_t len);
-
 int  vsnprintfz(char *dst, size_t n, const char *fmt, va_list args);
 int  snprintfz(char *dst, size_t n, const char *fmt, ...) PRINTFLIKE(3, 4);
-
-// memory allocation functions that handle failures
-#ifdef NETDATA_TRACE_ALLOCATIONS
-int malloc_trace_walkthrough(int (*callback)(void *item, void *data), void *data);
-
-#define strdupz(s) strdupz_int(s, __FILE__, __FUNCTION__, __LINE__)
-#define strndupz(s, len) strndupz_int(s, len, __FILE__, __FUNCTION__, __LINE__)
-#define callocz(nmemb, size) callocz_int(nmemb, size, __FILE__, __FUNCTION__, __LINE__)
-#define mallocz(size) mallocz_int(size, __FILE__, __FUNCTION__, __LINE__)
-#define reallocz(ptr, size) reallocz_int(ptr, size, __FILE__, __FUNCTION__, __LINE__)
-#define freez(ptr) freez_int(ptr, __FILE__, __FUNCTION__, __LINE__)
-#define mallocz_usable_size(ptr) mallocz_usable_size_int(ptr, __FILE__, __FUNCTION__, __LINE__)
-
-char *strdupz_int(const char *s, const char *file, const char *function, size_t line);
-char *strndupz_int(const char *s, size_t len, const char *file, const char *function, size_t line);
-void *callocz_int(size_t nmemb, size_t size, const char *file, const char *function, size_t line);
-void *mallocz_int(size_t size, const char *file, const char *function, size_t line);
-void *reallocz_int(void *ptr, size_t size, const char *file, const char *function, size_t line);
-void freez_int(void *ptr, const char *file, const char *function, size_t line);
-size_t mallocz_usable_size_int(void *ptr, const char *file, const char *function, size_t line);
-
-#else // NETDATA_TRACE_ALLOCATIONS
-char *strdupz(const char *s) MALLOCLIKE NEVERNULL;
-char *strndupz(const char *s, size_t len) MALLOCLIKE NEVERNULL;
-void *callocz(size_t nmemb, size_t size) MALLOCLIKE NEVERNULL;
-void *mallocz(size_t size) MALLOCLIKE NEVERNULL;
-void *reallocz(void *ptr, size_t size) MALLOCLIKE NEVERNULL;
-void freez(void *ptr);
-#endif // NETDATA_TRACE_ALLOCATIONS
-
-void mallocz_release_as_much_memory_to_the_system(void);
-void posix_memfree(void *ptr);
 
 void json_escape_string(char *dst, const char *src, size_t size);
 void json_fix_string(char *s);
 
-extern size_t netdata_mmap_count;
-void *netdata_mmap(const char *filename, size_t size, int flags, int ksm, bool read_only, bool dont_dump, int *open_fd);
-int netdata_munmap(void *ptr, size_t size);
-int memory_file_save(const char *filename, void *mem, size_t size);
 
 extern struct rlimit rlimit_nofile;
-
-extern int enable_ksm;
 
 char *fgets_trim_len(char *buf, size_t buf_size, FILE *fp, size_t *len);
 
@@ -187,43 +143,6 @@ extern const char *netdata_configured_host_prefix;
 #include "facets/facets.h"
 #include "functions_evloop/functions_evloop.h"
 #include "query_progress/progress.h"
-
-static inline size_t struct_natural_alignment(size_t size) __attribute__((const));
-
-#define STRUCT_NATURAL_ALIGNMENT (sizeof(uintptr_t) * 2)
-static inline size_t struct_natural_alignment(size_t size) {
-    if(unlikely(size % STRUCT_NATURAL_ALIGNMENT))
-        size = size + STRUCT_NATURAL_ALIGNMENT - (size % STRUCT_NATURAL_ALIGNMENT);
-
-    return size;
-}
-
-#ifdef NETDATA_TRACE_ALLOCATIONS
-struct malloc_trace {
-    avl_t avl;
-
-    const char *function;
-    const char *file;
-    size_t line;
-
-    size_t malloc_calls;
-    size_t calloc_calls;
-    size_t realloc_calls;
-    size_t strdup_calls;
-    size_t free_calls;
-
-    size_t mmap_calls;
-    size_t munmap_calls;
-
-    size_t allocations;
-    size_t bytes;
-
-    struct rrddim *rd_bytes;
-    struct rrddim *rd_allocations;
-    struct rrddim *rd_avg_alloc;
-    struct rrddim *rd_ops;
-};
-#endif // NETDATA_TRACE_ALLOCATIONS
 
 static inline PPvoid_t JudyLFirstThenNext(Pcvoid_t PArray, Word_t * PIndex, bool *first) {
     if(unlikely(*first)) {
